AI_Resume_Builder/requirements.txt
Flask
pdfplumber
python-docx
Werkzeug
